--- core/src/domain/aegis/policies.rs
@@ -1,99 +1,0 @@
-use crate::domain::{
-    aegis::ports::ClientScopePolicy,
-    authentication::value_objects::Identity,
-    client::ports::ClientRepository,
-    common::{
-        entities::app_errors::CoreError,
-        policies::{FerriskeyPolicy, Policy},
-    },
-    realm::entities::Realm,
-    role::entities::permission::Permissions,
-    user::ports::{UserRepository, UserRoleRepository},
-};
-
-impl<U, C, UR> ClientScopePolicy for FerriskeyPolicy<U, C, UR>
-where
-    U: UserRepository,
-    C: ClientRepository,
-    UR: UserRoleRepository,
-{
-    async fn can_create_scope(
-        &self,
-        identity: &Identity,
-        target_realm: &Realm,
-    ) -> Result<bool, CoreError> {
-        let user = self.get_user_from_identity(identity).await?;
-
-        let permissions = self
-            .get_permission_for_target_realm(&user, target_realm)
-            .await?;
-
-        let has_permission = Permissions::has_one_of_permissions(
-            &permissions,
-            &[Permissions::ManageRealm, Permissions::ManageClientScopes],
-        );
-
-        Ok(has_permission)
-    }
-
-    async fn can_view_scope(
-        &self,
-        identity: &Identity,
-        target_realm: &Realm,
-    ) -> Result<bool, CoreError> {
-        let user = self.get_user_from_identity(identity).await?;
-
-        let permissions = self
-            .get_permission_for_target_realm(&user, target_realm)
-            .await?;
-
-        let has_permission = Permissions::has_one_of_permissions(
-            &permissions,
-            &[
-                Permissions::ManageRealm,
-                Permissions::ManageClientScopes,
-                Permissions::ViewClientScopes,
-            ],
-        );
-
-        Ok(has_permission)
-    }
-
-    async fn can_update_scope(
-        &self,
-        identity: &Identity,
-        target_realm: &Realm,
-    ) -> Result<bool, CoreError> {
-        let user = self.get_user_from_identity(identity).await?;
-
-        let permissions = self
-            .get_permission_for_target_realm(&user, target_realm)
-            .await?;
-
-        let has_permission = Permissions::has_one_of_permissions(
-            &permissions,
-            &[Permissions::ManageRealm, Permissions::ManageClientScopes],
-        );
-
-        Ok(has_permission)
-    }
-
-    async fn can_delete_scope(
-        &self,
-        identity: &Identity,
-        target_realm: &Realm,
-    ) -> Result<bool, CoreError> {
-        let user = self.get_user_from_identity(identity).await?;
-
-        let permissions = self
-            .get_permission_for_target_realm(&user, target_realm)
-            .await?;
-
-        let has_permission = Permissions::has_one_of_permissions(
-            &permissions,
-            &[Permissions::ManageRealm, Permissions::ManageClientScopes],
-        );
-
-        Ok(has_permission)
-    }
-}
