+++ core/src/domain/organization/mod.rs
@@ -1,4 +1,3 @@
 pub mod group_services;
-pub mod policies;
 pub mod ports;
 pub mod services;
