+++ core/src/domain/realm/mod.rs
@@ -1,4 +1,3 @@
 pub mod entities;
-pub mod policies;
 pub mod ports;
 pub mod services;
