+++ core/src/domain/user/mod.rs
@@ -1,5 +1,4 @@
 pub mod entities;
-pub mod policies;
 pub mod ports;
 pub mod services;
 pub mod value_objects;
