+++ libs/ferriskey-abyss/src/identity_provider/mod.rs
@@ -10,3 +10,5 @@ pub use entities::{
 };
 pub use ports::{IdentityProviderPolicy, IdentityProviderRepository, IdentityProviderService};
 pub use value_objects::{CreateIdentityProviderRequest, UpdateIdentityProviderRequest};
+
+pub mod policies;
