+++ libs/ferriskey-aegis/src/lib.rs
@@ -1,3 +1,5 @@
 pub mod entities;
 pub mod ports;
 pub mod value_objects;
+
+pub mod policies;
