+++ libs/ferriskey-aegis/src/policies.rs
@@ -1,22 +1,20 @@
-use crate::domain::{
-    authentication::value_objects::Identity,
-    client::ports::ClientRepository,
-    common::{
-        entities::app_errors::CoreError,
-        policies::{FerriskeyPolicy, Policy},
-    },
-    realm::entities::Realm,
-    role::entities::permission::Permissions,
-    user::ports::{UserPolicy, UserRepository, UserRoleRepository},
-};
+use ferriskey_domain::auth::Identity;
+use ferriskey_domain::client::ports::ClientRepository;
+use ferriskey_domain::common::app_errors::CoreError;
+use ferriskey_domain::common::policies::{FerriskeyPolicy, Policy};
+use ferriskey_domain::realm::Realm;
+use ferriskey_domain::role::permission::Permissions;
+use ferriskey_domain::user::ports::{UserRepository, UserRoleRepository};
 
-impl<U, C, UR> UserPolicy for FerriskeyPolicy<U, C, UR>
+use crate::ports::ClientScopePolicy;
+
+impl<U, C, UR> ClientScopePolicy for FerriskeyPolicy<U, C, UR>
 where
     U: UserRepository,
     C: ClientRepository,
     UR: UserRoleRepository,
 {
-    async fn can_create_user(
+    async fn can_create_scope(
         &self,
         identity: &Identity,
         target_realm: &Realm,
@@ -29,80 +27,18 @@ where
 
         let has_permission = Permissions::has_one_of_permissions(
             &permissions,
-            &[Permissions::ManageRealm, Permissions::ManageUsers],
+            &[Permissions::ManageRealm, Permissions::ManageClientScopes],
         );
 
         Ok(has_permission)
     }
 
-    async fn can_delete_user(
+    async fn can_view_scope(
         &self,
         identity: &Identity,
         target_realm: &Realm,
     ) -> Result<bool, CoreError> {
         let user = self.get_user_from_identity(identity).await?;
-
-        let permissions = self
-            .get_permission_for_target_realm(&user, target_realm)
-            .await?;
-
-        let has_permission = Permissions::has_one_of_permissions(
-            &permissions,
-            &[Permissions::ManageRealm, Permissions::ManageUsers],
-        );
-
-        Ok(has_permission)
-    }
-
-    async fn can_update_user(
-        &self,
-        identity: &Identity,
-        target_realm: &Realm,
-    ) -> Result<bool, CoreError> {
-        let user = self.get_user_from_identity(identity).await?;
-
-        let permissions = self
-            .get_permission_for_target_realm(&user, target_realm)
-            .await?;
-
-        let has_permission = Permissions::has_one_of_permissions(
-            &permissions,
-            &[Permissions::ManageRealm, Permissions::ManageUsers],
-        );
-
-        Ok(has_permission)
-    }
-
-    async fn can_view_user(
-        &self,
-        identity: &Identity,
-        target_realm: &Realm,
-    ) -> Result<bool, CoreError> {
-        let user = self.get_user_from_identity(identity).await?;
-
-        let permissions = self
-            .get_permission_for_target_realm(&user, target_realm)
-            .await?;
-
-        let has_permission = Permissions::has_one_of_permissions(
-            &permissions,
-            &[Permissions::ManageRealm, Permissions::ViewUsers],
-        );
-
-        Ok(has_permission)
-    }
-
-    async fn can_view_user_permissions(
-        &self,
-        identity: &Identity,
-        target_realm: &Realm,
-        target_user_id: uuid::Uuid,
-    ) -> Result<bool, CoreError> {
-        let user = self.get_user_from_identity(identity).await?;
-
-        if user.id == target_user_id {
-            return Ok(true);
-        }
 
         let permissions = self
             .get_permission_for_target_realm(&user, target_realm)
@@ -112,9 +48,47 @@ where
             &permissions,
             &[
                 Permissions::ManageRealm,
-                Permissions::ManageUsers,
-                Permissions::ViewUsers,
+                Permissions::ManageClientScopes,
+                Permissions::ViewClientScopes,
             ],
+        );
+
+        Ok(has_permission)
+    }
+
+    async fn can_update_scope(
+        &self,
+        identity: &Identity,
+        target_realm: &Realm,
+    ) -> Result<bool, CoreError> {
+        let user = self.get_user_from_identity(identity).await?;
+
+        let permissions = self
+            .get_permission_for_target_realm(&user, target_realm)
+            .await?;
+
+        let has_permission = Permissions::has_one_of_permissions(
+            &permissions,
+            &[Permissions::ManageRealm, Permissions::ManageClientScopes],
+        );
+
+        Ok(has_permission)
+    }
+
+    async fn can_delete_scope(
+        &self,
+        identity: &Identity,
+        target_realm: &Realm,
+    ) -> Result<bool, CoreError> {
+        let user = self.get_user_from_identity(identity).await?;
+
+        let permissions = self
+            .get_permission_for_target_realm(&user, target_realm)
+            .await?;
+
+        let has_permission = Permissions::has_one_of_permissions(
+            &permissions,
+            &[Permissions::ManageRealm, Permissions::ManageClientScopes],
         );
 
         Ok(has_permission)
