+++ libs/ferriskey-compass/src/lib.rs
@@ -2,3 +2,5 @@ pub mod entities;
 pub mod ports;
 pub mod recorder;
 pub mod value_objects;
+
+pub mod policies;
