+++ libs/ferriskey-organization/src/lib.rs
@@ -3,3 +3,5 @@ pub mod ports;
 
 pub use entities::*;
 pub use ports::*;
+
+pub mod policies;
